+++ src/commands/exec.ts
@@ -1,8 +1,8 @@
 import { commands } from 'vscode'
-import { handler } from '../menu/exec'
+import { run } from '../menu/exec'
 
 export function exec() {
   return commands.registerCommand('argon.exec', () => {
-    handler()
+    run()
   })
 }