+++ src/menu/build.ts
@@ -3,7 +3,7 @@ import * as argon from '../argon'
 import { getProjectName } from '../util'
 import { Item, getProject } from '.'
 import { State } from '../state'
-import { Session } from '../session'
+import { RestorableSession, Session } from '../session'
 
 export const item: Item = {
   label: '$(file-binary) Build',
@@ -11,8 +11,43 @@ export const item: Item = {
   action: 'build',
 }
 
-function getOutput(name: string): Promise<string | undefined> {
+const OPTIONS = [
+  {
+    label: 'Watch for changes',
+    flag: '--watch',
+    picked: true,
+  },
+  {
+    label: 'Generate sourcemap',
+    flag: '--sourcemap',
+    picked: true,
+  },
+  {
+    label: 'Build plugin',
+    flag: '--plugin',
+    picked: false,
+  },
+  {
+    label: 'Use XML format',
+    flag: '--xml',
+    picked: false,
+  },
+  {
+    label: 'Use roblox-ts',
+    flag: '--ts',
+    picked: false,
+  },
+]
+
+function getOutput(
+  name: string,
+  restore: boolean,
+): Promise<string | undefined> {
   return new Promise((resolve, reject) => {
+    if (restore) {
+      return resolve(undefined)
+    }
+
     vscode.window
       .showInputBox({
         title: 'Enter build output',
@@ -29,47 +64,26 @@ function getOutput(name: string): Promise<string | undefined> {
   })
 }
 
-function getOptions(context: vscode.ExtensionContext): Promise<string[]> {
+function getOptions(
+  context: vscode.ExtensionContext,
+  restore: boolean,
+): Promise<string[]> {
   return new Promise((resolve, reject) => {
-    const options = [
-      {
-        label: 'Watch for changes',
-        flag: '--watch',
-        id: 'watch',
-        picked: true,
-      },
-      {
-        label: 'Generate sourcemap',
-        flag: '--sourcemap',
-        id: 'sourcemap',
-        picked: true,
-      },
-      {
-        label: 'Build plugin',
-        flag: '--plugin',
-        id: 'plugin',
-        picked: false,
-      },
-      {
-        label: 'Use XML format',
-        flag: '--xml',
-        id: 'xml',
-        picked: false,
-      },
-      {
-        label: 'Use roblox-ts',
-        flag: '--ts',
-        id: 'tsMode',
-        picked: false,
-      },
-    ]
-
-    options.forEach((option) => {
-      option['picked'] = context.globalState.get(option.id, option.picked)
+    OPTIONS.forEach((option) => {
+      option['picked'] = context.globalState.get(
+        'Build' + option.flag,
+        option.picked,
+      )
     })
 
+    if (restore) {
+      return resolve(
+        OPTIONS.flatMap((option) => (option.picked ? [option.flag] : [])),
+      )
+    }
+
     vscode.window
-      .showQuickPick(options, {
+      .showQuickPick(OPTIONS, {
         title: 'Select build options',
         canPickMany: true,
       })
@@ -78,10 +92,10 @@ function getOptions(context: vscode.ExtensionContext): Promise<string[]> {
           return reject()
         }
 
-        options.forEach((item) => {
+        OPTIONS.forEach((item) => {
           context.globalState.update(
-            item.id,
-            items.some((i) => i.id === item.id),
+            'Build' + item.flag,
+            items.some((i) => i.flag === item.flag),
           )
         })
 
@@ -90,13 +104,20 @@ function getOptions(context: vscode.ExtensionContext): Promise<string[]> {
   })
 }
 
-export async function handler(state: State) {
-  const project = await getProject(state.context)
+export async function run(state: State, session?: RestorableSession) {
+  const restore = !!session
+
+  console.log(restore)
+
+  if (!restore) {
+    var project = await getProject(state.context)
+  } else {
+    var project = session.project
+  }
 
   const name = getProjectName(project)
-  const output = await getOutput(name)
-
-  const options = await getOptions(state.context)
+  const output = await getOutput(name, restore)
+  const options = await getOptions(state.context, restore)
 
   if (output) {
     options.push('--output', output)
@@ -105,7 +126,6 @@ export async function handler(state: State) {
   const id = await argon.build(project, options)
 
   if (id) {
-    const session = new Session(name, project, id).withType('Build')
-    state.addSession(session)
+    state.addSession(new Session(name, project, id).withType('Build'))
   }
 }