+++ src/menu/debug.ts
@@ -47,7 +47,7 @@ function getMode(): Promise<string> {
   })
 }
 
-export async function handler() {
+export async function run() {
   const mode = await getMode()
 
   argon.debug(mode)