+++ src/menu/plugin.ts
@@ -35,7 +35,7 @@ function getMode(): Promise<string> {
   })
 }
 
-export async function handler() {
+export async function run() {
   const mode = await getMode()
 
   argon.plugin(mode)