+++ src/session.ts
@@ -5,7 +5,7 @@ export class Session {
   public name: string
   public project: string
   public id: number
-  public address: string | undefined
+  public address?: string
 
   public constructor(name: string, project: string, id: number) {
     this.name = name
@@ -25,5 +25,37 @@ export class Session {
 
   public equals(other: Session) {
     return this.type === other.type && this.project === other.project
+  }
+}
+
+export class RestorableSession {
+  public type: SessionType
+  public project: string
+  public address?: string
+
+  private isComplete: boolean = true
+
+  public constructor(session: Session | any) {
+    if (session instanceof Session) {
+      this.type = session.type
+      this.project = session.project
+      this.address = session.address
+    } else {
+      if (!session || !session.type || !session.project) {
+        this.isComplete = false
+        this.type = 'Serve'
+        this.project = ''
+
+        return
+      }
+
+      this.type = session.type
+      this.project = session.project
+      this.address = session.address
+    }
+  }
+
+  public isRestorable() {
+    return this.isComplete
   }
 }